pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multiplex-image-model"
version = "0.1.0"
description = "A masked autoencoder for multiplex imaging data with hyperkernel architecture"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["multiplex imaging", "autoencoder", "hyperkernel", "deep learning"]
classifiers = [
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
]

dependencies = [
    "tifffile==2025.6.11",
    "torch==2.7.1",
    "numpy==2.2.6",
    "torchvision==0.22.1",
    "tqdm==4.67.1",
    "comet-ml>=3.0.0",
    "pydantic>=2.0.0",
    "matplotlib==3.10.3",
    "pandas==2.3.1",
    "scikit-image==0.25.2",
    "scikit-learn==1.7.1",
    "seaborn==0.13.2",
    "ruamel.yaml==0.18.14",
    "opencv-python==4.12.0.88",
    "gpytorch==1.15.1"
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/dav3794/multiplex-image-model"
Repository = "https://github.com/dav3794/multiplex-image-model"
Issues = "https://github.com/dav3794/multiplex-image-model/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["multiplex_model*"]
exclude = ["tests*", "notebooks*", "configs*"]

[tool.black]
line-length = 120
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true
explicit_package_bases = true

[[tool.mypy.overrides]]
# multiplex_model.utils is fully typed — enforce it stays that way
module = "multiplex_model.utils.*"
disallow_untyped_defs = true
check_untyped_defs = true
